refactor: extract Research Management Service from GameManager
Extract research-related operations into ResearchManagementService to continue
GameManager decomposition following the established refactoring plan.

Changes:
- Create ResearchManagementService with extracted research operations
  - setPlayerResearch(), setResearchGoal()
  - getPlayerResearch(), getAvailableTechnologies(), getResearchProgress()
  - processResearchTurn() with tech completion handling
- Update GameManager to delegate to ResearchManagementService
- Register service in ServiceRegistry for proper lifecycle management
- Maintain all existing functionality and broadcasting patterns

Progress:
- GameManager reduced from 1,634 to 1,557 lines (77 line reduction)
- Total progress: 507 lines extracted from original 2,064 lines (25% reduction)
